chore: move utils functions to helper files
