+++ src/helpers/arrayEquals.js
@@ -1,0 +1,5 @@
+import { isEmpty, isEqual, xorWith } from 'lodash'
+
+const arrayEquals = (x, y) => isEmpty(xorWith(x, y, isEqual))
+
+export default arrayEquals
