+++ src/helpers/groupResult.js
@@ -1,0 +1,48 @@
+import isGroupable from './isGroupable'
+
+const groupResult = (data, groupBy, key = 'id') => {
+  // Transform a list of records to a TreeTable-compatible structure
+  // with grouped records.
+  let result = []
+  let existingGroups = []
+  for (const item of data) {
+    // Unique items. Just add to root
+    if (!isGroupable(data, groupBy, item[groupBy])) {
+      result.push({ key: item[key], data: item })
+    }
+
+    // Item of an existing group
+    else if (existingGroups.includes(item[groupBy])) {
+      for (const group of result) {
+        if (group.data[groupBy] === item[groupBy]) {
+          group.children.push({
+            key: item[key],
+            data: item,
+          })
+          break
+        }
+      }
+    }
+
+    // New group
+    else {
+      existingGroups.push(item[groupBy])
+      result.push({
+        key: `group-${item[groupBy]}`,
+        data: {
+          [groupBy]: item[groupBy],
+          isGroup: true,
+        },
+        children: [
+          {
+            key: item[key],
+            data: item,
+          },
+        ],
+      })
+    }
+  }
+  return result
+}
+
+export default groupResult
